Make dynamic allocation heuristics pluggable
