+++ core/src/main/scala/org/apache/spark/ExecutorAllocationManager.scala
@@ -24,57 +24,22 @@ import scala.util.control.ControlThrowable
 
 import com.codahale.metrics.{Gauge, MetricRegistry}
 
+import org.apache.spark.dynamicallocation.{AllocationStrategy, DefaultAllocationStrategy, Metric}
 import org.apache.spark.internal.Logging
 import org.apache.spark.metrics.source.Source
 import org.apache.spark.scheduler._
-import org.apache.spark.util.{Clock, SystemClock, ThreadUtils}
+import org.apache.spark.util.{Clock, SystemClock, ThreadUtils, Utils}
 
 /**
  * An agent that dynamically allocates and removes executors based on the workload.
  *
- * The ExecutorAllocationManager maintains a moving target number of executors which is periodically
- * synced to the cluster manager. The target starts at a configured initial value and changes with
- * the number of pending and running tasks.
- *
- * Decreasing the target number of executors happens when the current target is more than needed to
- * handle the current load. The target number of executors is always truncated to the number of
- * executors that could run all current running and pending tasks at once.
- *
- * Increasing the target number of executors happens in response to backlogged tasks waiting to be
- * scheduled. If the scheduler queue is not drained in N seconds, then new executors are added. If
- * the queue persists for another M seconds, then more executors are added and so on. The number
- * added in each round increases exponentially from the previous round until an upper bound has been
- * reached. The upper bound is based both on a configured property and on the current number of
- * running and pending tasks, as described above.
- *
- * The rationale for the exponential increase is twofold: (1) Executors should be added slowly
- * in the beginning in case the number of extra executors needed turns out to be small. Otherwise,
- * we may add more executors than we need just to remove them later. (2) Executors should be added
- * quickly over time in case the maximum number of executors is very high. Otherwise, it will take
- * a long time to ramp up under heavy workloads.
- *
- * The remove policy is simpler: If an executor has been idle for K seconds, meaning it has not
- * been scheduled to run any tasks, then it is removed.
- *
- * There is no retry logic in either case because we make the assumption that the cluster manager
- * will eventually fulfill all requests it receives asynchronously.
+ * The ExecutorAllocationManager periodically calls the configured dynamic allocation strategy to
+ * perform the actual executor allocation/release.
  *
  * The relevant Spark properties include the following:
  *
  *   spark.dynamicAllocation.enabled - Whether this feature is enabled
- *   spark.dynamicAllocation.minExecutors - Lower bound on the number of executors
- *   spark.dynamicAllocation.maxExecutors - Upper bound on the number of executors
- *   spark.dynamicAllocation.initialExecutors - Number of executors to start with
- *
- *   spark.dynamicAllocation.schedulerBacklogTimeout (M) -
- *     If there are backlogged tasks for this duration, add new executors
- *
- *   spark.dynamicAllocation.sustainedSchedulerBacklogTimeout (N) -
- *     If the backlog is sustained for this duration, add more executors
- *     This is used only after the initial backlog timeout is exceeded
- *
- *   spark.dynamicAllocation.executorIdleTimeout (K) -
- *     If an executor has been idle for this duration, remove it
+ *   spark.dynamicAllocation.strategy - The class name of the dynamic allocation strategy
  */
 private[spark] class ExecutorAllocationManager(
     client: ExecutorAllocationClient,
@@ -84,61 +49,26 @@ private[spark] class ExecutorAllocationManager(
 
   allocationManager =>
 
-  import ExecutorAllocationManager._
+  // strategy to use for dynamic allocation
+  private val allocationStrategyClass = conf.get("spark.dynamicAllocation.strategy",
+    classOf[DefaultAllocationStrategy].getName)
 
-  // Lower and upper bounds on the number of executors.
-  private val minNumExecutors = conf.getInt("spark.dynamicAllocation.minExecutors", 0)
-  private val maxNumExecutors = conf.getInt("spark.dynamicAllocation.maxExecutors",
-    Integer.MAX_VALUE)
+  private[spark] lazy val allocationStrategy = try {
+    Utils.classForName(allocationStrategyClass)
+      .getDeclaredConstructor(classOf[SparkConf],
+        classOf[ExecutorAllocationManager], classOf[ExecutorAllocationClient])
+      .newInstance(conf, this, client)
+      .asInstanceOf[AllocationStrategy]
+  } catch {
+    case e: Exception =>
+      throw new SparkException(s"Unable to load ${allocationStrategyClass}", e)
+  }
+
   private val initialNumExecutors = conf.getInt("spark.dynamicAllocation.initialExecutors",
-    minNumExecutors)
-
-  // How long there must be backlogged tasks for before an addition is triggered (seconds)
-  private val schedulerBacklogTimeoutS = conf.getTimeAsSeconds(
-    "spark.dynamicAllocation.schedulerBacklogTimeout", "1s")
-
-  // Same as above, but used only after `schedulerBacklogTimeoutS` is exceeded
-  private val sustainedSchedulerBacklogTimeoutS = conf.getTimeAsSeconds(
-    "spark.dynamicAllocation.sustainedSchedulerBacklogTimeout", s"${schedulerBacklogTimeoutS}s")
-
-  // How long an executor must be idle for before it is removed (seconds)
-  private val executorIdleTimeoutS = conf.getTimeAsSeconds(
-    "spark.dynamicAllocation.executorIdleTimeout", "60s")
-
-  private val cachedExecutorIdleTimeoutS = conf.getTimeAsSeconds(
-    "spark.dynamicAllocation.cachedExecutorIdleTimeout", s"${Integer.MAX_VALUE}s")
-
-  // During testing, the methods to actually kill and add executors are mocked out
-  private val testing = conf.getBoolean("spark.dynamicAllocation.testing", false)
-
-  // TODO: The default value of 1 for spark.executor.cores works right now because dynamic
-  // allocation is only supported for YARN and the default number of cores per executor in YARN is
-  // 1, but it might need to be attained differently for different cluster managers
-  private val tasksPerExecutor =
-    conf.getInt("spark.executor.cores", 1) / conf.getInt("spark.task.cpus", 1)
-
-  validateSettings()
-
-  // Number of executors to add in the next round
-  private var numExecutorsToAdd = 1
-
-  // The desired number of executors at this moment in time. If all our executors were to die, this
-  // is the number of executors we would immediately want from the cluster manager.
-  private var numExecutorsTarget = initialNumExecutors
-
-  // Executors that have been requested to be removed but have not been killed yet
-  private val executorsPendingToRemove = new mutable.HashSet[String]
+    conf.getInt("spark.dynamicAllocation.minExecutors", 0))
 
   // All known executors
-  private val executorIds = new mutable.HashSet[String]
-
-  // A timestamp of when an addition should be triggered, or NOT_SET if it is not set
-  // This is set when pending tasks are added but not scheduled yet
-  private var addTime: Long = NOT_SET
-
-  // A timestamp for each executor of when the executor should be removed, indexed by the ID
-  // This is set when an executor is no longer running a task, or when it first registers
-  private val removeTimes = new mutable.HashMap[String, Long]
+  private[spark] val executorIds = new mutable.HashSet[String]
 
   // Polling loop interval (ms)
   private val intervalMillis: Long = 100
@@ -147,7 +77,7 @@ private[spark] class ExecutorAllocationManager(
   private var clock: Clock = new SystemClock()
 
   // Listener for Spark events that impact the allocation policy
-  private val listener = new ExecutorAllocationListener
+  private[spark] val listener = new ExecutorAllocationListener
 
   // Executor that handles the scheduling task.
   private val executor =
@@ -161,55 +91,20 @@ private[spark] class ExecutorAllocationManager(
   // set to false when:
   //   (1) a stage is submitted, or
   //   (2) an executor idle timeout has elapsed.
-  @volatile private var initializing: Boolean = true
+  @volatile private[spark] var initializing: Boolean = true
 
   // Number of locality aware tasks, used for executor placement.
-  private var localityAwareTasks = 0
+  private[spark] var localityAwareTasks = 0
 
   // Host to possible task running on it, used for executor placement.
-  private var hostToLocalTaskCount: Map[String, Int] = Map.empty
-
-  /**
-   * Verify that the settings specified through the config are valid.
-   * If not, throw an appropriate exception.
-   */
-  private def validateSettings(): Unit = {
-    if (minNumExecutors < 0 || maxNumExecutors < 0) {
-      throw new SparkException("spark.dynamicAllocation.{min/max}Executors must be positive!")
-    }
-    if (maxNumExecutors == 0) {
-      throw new SparkException("spark.dynamicAllocation.maxExecutors cannot be 0!")
-    }
-    if (minNumExecutors > maxNumExecutors) {
-      throw new SparkException(s"spark.dynamicAllocation.minExecutors ($minNumExecutors) must " +
-        s"be less than or equal to spark.dynamicAllocation.maxExecutors ($maxNumExecutors)!")
-    }
-    if (schedulerBacklogTimeoutS <= 0) {
-      throw new SparkException("spark.dynamicAllocation.schedulerBacklogTimeout must be > 0!")
-    }
-    if (sustainedSchedulerBacklogTimeoutS <= 0) {
-      throw new SparkException(
-        "spark.dynamicAllocation.sustainedSchedulerBacklogTimeout must be > 0!")
-    }
-    if (executorIdleTimeoutS <= 0) {
-      throw new SparkException("spark.dynamicAllocation.executorIdleTimeout must be > 0!")
-    }
-    // Require external shuffle service for dynamic allocation
-    // Otherwise, we may lose shuffle files when killing executors
-    if (!conf.getBoolean("spark.shuffle.service.enabled", false) && !testing) {
-      throw new SparkException("Dynamic allocation of executors requires the external " +
-        "shuffle service. You may enable this through spark.shuffle.service.enabled.")
-    }
-    if (tasksPerExecutor == 0) {
-      throw new SparkException("spark.executor.cores must not be less than spark.task.cpus.")
-    }
-  }
+  private[spark] var hostToLocalTaskCount: Map[String, Int] = Map.empty
 
   /**
    * Use a different clock for this allocation manager. This is mainly used for testing.
    */
   def setClock(newClock: Clock): Unit = {
     clock = newClock
+    allocationStrategy.setClock(newClock)
   }
 
   /**
@@ -233,7 +128,7 @@ private[spark] class ExecutorAllocationManager(
     }
     executor.scheduleAtFixedRate(scheduleTask, 0, intervalMillis, TimeUnit.MILLISECONDS)
 
-    client.requestTotalExecutors(numExecutorsTarget, localityAwareTasks, hostToLocalTaskCount)
+    client.requestTotalExecutors(initialNumExecutors, localityAwareTasks, hostToLocalTaskCount)
   }
 
   /**
@@ -250,185 +145,23 @@ private[spark] class ExecutorAllocationManager(
    */
   def reset(): Unit = synchronized {
     initializing = true
-    numExecutorsTarget = initialNumExecutors
-    numExecutorsToAdd = 1
-
-    executorsPendingToRemove.clear()
-    removeTimes.clear()
-  }
-
-  /**
-   * The maximum number of executors we would need under the current load to satisfy all running
-   * and pending tasks, rounded up.
-   */
-  private def maxNumExecutorsNeeded(): Int = {
-    val numRunningOrPendingTasks = listener.totalPendingTasks + listener.totalRunningTasks
-    (numRunningOrPendingTasks + tasksPerExecutor - 1) / tasksPerExecutor
+    allocationStrategy.reset()
   }
 
   /**
    * This is called at a fixed interval to regulate the number of pending executor requests
    * and number of executors running.
    *
-   * First, adjust our requested executors based on the add time and our current needs.
-   * Then, if the remove time for an existing executor has expired, kill the executor.
+   * First, call allocation strategy's allocate() to get new executors if needed.
+   * Then, call release() remove the executors that are no longer needed.
    *
    * This is factored out into its own method for testing.
    */
   private def schedule(): Unit = synchronized {
     val now = clock.getTimeMillis
-
-    updateAndSyncNumExecutorsTarget(now)
-
-    removeTimes.retain { case (executorId, expireTime) =>
-      val expired = now >= expireTime
-      if (expired) {
-        initializing = false
-        removeExecutor(executorId)
-      }
-      !expired
-    }
-  }
-
-  /**
-   * Updates our target number of executors and syncs the result with the cluster manager.
-   *
-   * Check to see whether our existing allocation and the requests we've made previously exceed our
-   * current needs. If so, truncate our target and let the cluster manager know so that it can
-   * cancel pending requests that are unneeded.
-   *
-   * If not, and the add time has expired, see if we can request new executors and refresh the add
-   * time.
-   *
-   * @return the delta in the target number of executors.
-   */
-  private def updateAndSyncNumExecutorsTarget(now: Long): Int = synchronized {
-    val maxNeeded = maxNumExecutorsNeeded
-
-    if (initializing) {
-      // Do not change our target while we are still initializing,
-      // Otherwise the first job may have to ramp up unnecessarily
-      0
-    } else if (maxNeeded < numExecutorsTarget) {
-      // The target number exceeds the number we actually need, so stop adding new
-      // executors and inform the cluster manager to cancel the extra pending requests
-      val oldNumExecutorsTarget = numExecutorsTarget
-      numExecutorsTarget = math.max(maxNeeded, minNumExecutors)
-      numExecutorsToAdd = 1
-
-      // If the new target has not changed, avoid sending a message to the cluster manager
-      if (numExecutorsTarget < oldNumExecutorsTarget) {
-        client.requestTotalExecutors(numExecutorsTarget, localityAwareTasks, hostToLocalTaskCount)
-        logDebug(s"Lowering target number of executors to $numExecutorsTarget (previously " +
-          s"$oldNumExecutorsTarget) because not all requested executors are actually needed")
-      }
-      numExecutorsTarget - oldNumExecutorsTarget
-    } else if (addTime != NOT_SET && now >= addTime) {
-      val delta = addExecutors(maxNeeded)
-      logDebug(s"Starting timer to add more executors (to " +
-        s"expire in $sustainedSchedulerBacklogTimeoutS seconds)")
-      addTime += sustainedSchedulerBacklogTimeoutS * 1000
-      delta
-    } else {
-      0
-    }
-  }
-
-  /**
-   * Request a number of executors from the cluster manager.
-   * If the cap on the number of executors is reached, give up and reset the
-   * number of executors to add next round instead of continuing to double it.
-   *
-   * @param maxNumExecutorsNeeded the maximum number of executors all currently running or pending
-   *                              tasks could fill
-   * @return the number of additional executors actually requested.
-   */
-  private def addExecutors(maxNumExecutorsNeeded: Int): Int = {
-    // Do not request more executors if it would put our target over the upper bound
-    if (numExecutorsTarget >= maxNumExecutors) {
-      logDebug(s"Not adding executors because our current target total " +
-        s"is already $numExecutorsTarget (limit $maxNumExecutors)")
-      numExecutorsToAdd = 1
-      return 0
-    }
-
-    val oldNumExecutorsTarget = numExecutorsTarget
-    // There's no point in wasting time ramping up to the number of executors we already have, so
-    // make sure our target is at least as much as our current allocation:
-    numExecutorsTarget = math.max(numExecutorsTarget, executorIds.size)
-    // Boost our target with the number to add for this round:
-    numExecutorsTarget += numExecutorsToAdd
-    // Ensure that our target doesn't exceed what we need at the present moment:
-    numExecutorsTarget = math.min(numExecutorsTarget, maxNumExecutorsNeeded)
-    // Ensure that our target fits within configured bounds:
-    numExecutorsTarget = math.max(math.min(numExecutorsTarget, maxNumExecutors), minNumExecutors)
-
-    val delta = numExecutorsTarget - oldNumExecutorsTarget
-
-    // If our target has not changed, do not send a message
-    // to the cluster manager and reset our exponential growth
-    if (delta == 0) {
-      numExecutorsToAdd = 1
-      return 0
-    }
-
-    val addRequestAcknowledged = testing ||
-      client.requestTotalExecutors(numExecutorsTarget, localityAwareTasks, hostToLocalTaskCount)
-    if (addRequestAcknowledged) {
-      val executorsString = "executor" + { if (delta > 1) "s" else "" }
-      logInfo(s"Requesting $delta new $executorsString because tasks are backlogged" +
-        s" (new desired total will be $numExecutorsTarget)")
-      numExecutorsToAdd = if (delta == numExecutorsToAdd) {
-        numExecutorsToAdd * 2
-      } else {
-        1
-      }
-      delta
-    } else {
-      logWarning(
-        s"Unable to reach the cluster manager to request $numExecutorsTarget total executors!")
-      numExecutorsTarget = oldNumExecutorsTarget
-      0
-    }
-  }
-
-  /**
-   * Request the cluster manager to remove the given executor.
-   * Return whether the request is received.
-   */
-  private def removeExecutor(executorId: String): Boolean = synchronized {
-    // Do not kill the executor if we are not aware of it (should never happen)
-    if (!executorIds.contains(executorId)) {
-      logWarning(s"Attempted to remove unknown executor $executorId!")
-      return false
-    }
-
-    // Do not kill the executor again if it is already pending to be killed (should never happen)
-    if (executorsPendingToRemove.contains(executorId)) {
-      logWarning(s"Attempted to remove executor $executorId " +
-        s"when it is already pending to be removed!")
-      return false
-    }
-
-    // Do not kill the executor if we have already reached the lower bound
-    val numExistingExecutors = executorIds.size - executorsPendingToRemove.size
-    if (numExistingExecutors - 1 < minNumExecutors) {
-      logDebug(s"Not removing idle executor $executorId because there are only " +
-        s"$numExistingExecutors executor(s) left (limit $minNumExecutors)")
-      return false
-    }
-
-    // Send a request to the backend to kill this executor
-    val removeRequestAcknowledged = testing || client.killExecutor(executorId)
-    if (removeRequestAcknowledged) {
-      logInfo(s"Removing executor $executorId because it has been idle for " +
-        s"$executorIdleTimeoutS seconds (new desired total will be ${numExistingExecutors - 1})")
-      executorsPendingToRemove.add(executorId)
-      true
-    } else {
-      logWarning(s"Unable to reach the cluster manager to kill executor $executorId," +
-        s"or no executor eligible to kill!")
-      false
+    allocationStrategy.allocate(now)
+    if (allocationStrategy.release(now)) {
+      initializing = false
     }
   }
 
@@ -447,6 +180,7 @@ private[spark] class ExecutorAllocationManager(
     } else {
       logWarning(s"Duplicate executor $executorId has registered")
     }
+    allocationStrategy.onExecutorAdded(executorId)
   }
 
   /**
@@ -455,13 +189,8 @@ private[spark] class ExecutorAllocationManager(
   private def onExecutorRemoved(executorId: String): Unit = synchronized {
     if (executorIds.contains(executorId)) {
       executorIds.remove(executorId)
-      removeTimes.remove(executorId)
       logInfo(s"Existing executor $executorId has been removed (new total is ${executorIds.size})")
-      if (executorsPendingToRemove.contains(executorId)) {
-        executorsPendingToRemove.remove(executorId)
-        logDebug(s"Executor $executorId is no longer pending to " +
-          s"be removed (${executorsPendingToRemove.size} left)")
-      }
+      allocationStrategy.onExecutorRemoved(executorId)
     } else {
       logWarning(s"Unknown executor $executorId has been removed!")
     }
@@ -473,11 +202,7 @@ private[spark] class ExecutorAllocationManager(
    * if it is not already set.
    */
   private def onSchedulerBacklogged(): Unit = synchronized {
-    if (addTime == NOT_SET) {
-      logDebug(s"Starting timer to add executors because pending tasks " +
-        s"are building up (to expire in $schedulerBacklogTimeoutS seconds)")
-      addTime = clock.getTimeMillis + schedulerBacklogTimeoutS * 1000
-    }
+    allocationStrategy.onSchedulerBacklogged()
   }
 
   /**
@@ -485,9 +210,7 @@ private[spark] class ExecutorAllocationManager(
    * This resets all variables used for adding executors.
    */
   private def onSchedulerQueueEmpty(): Unit = synchronized {
-    logDebug("Clearing timer to add executors because there are no more pending tasks")
-    addTime = NOT_SET
-    numExecutorsToAdd = 1
+    allocationStrategy.onSchedulerQueueEmpty()
   }
 
   /**
@@ -497,25 +220,7 @@ private[spark] class ExecutorAllocationManager(
    */
   private def onExecutorIdle(executorId: String): Unit = synchronized {
     if (executorIds.contains(executorId)) {
-      if (!removeTimes.contains(executorId) && !executorsPendingToRemove.contains(executorId)) {
-        // Note that it is not necessary to query the executors since all the cached
-        // blocks we are concerned with are reported to the driver. Note that this
-        // does not include broadcast blocks.
-        val hasCachedBlocks = SparkEnv.get.blockManager.master.hasCachedBlocks(executorId)
-        val now = clock.getTimeMillis()
-        val timeout = {
-          if (hasCachedBlocks) {
-            // Use a different timeout if the executor has cached blocks.
-            now + cachedExecutorIdleTimeoutS * 1000
-          } else {
-            now + executorIdleTimeoutS * 1000
-          }
-        }
-        val realTimeout = if (timeout <= 0) Long.MaxValue else timeout // overflow
-        removeTimes(executorId) = realTimeout
-        logDebug(s"Starting idle timer for $executorId because there are no more tasks " +
-          s"scheduled to run on the executor (to expire in ${(realTimeout - now)/1000} seconds)")
-      }
+      allocationStrategy.onExecutorIdle(executorId)
     } else {
       logWarning(s"Attempted to mark unknown executor $executorId idle")
     }
@@ -526,8 +231,7 @@ private[spark] class ExecutorAllocationManager(
    * This resets all variables used for removing this executor.
    */
   private def onExecutorBusy(executorId: String): Unit = synchronized {
-    logDebug(s"Clearing idle timer for $executorId because it is now running a task")
-    removeTimes.remove(executorId)
+    allocationStrategy.onExecutorBusy(executorId)
   }
 
   /**
@@ -537,7 +241,7 @@ private[spark] class ExecutorAllocationManager(
    * and consistency of events returned by the listener. For simplicity, it does not account
    * for speculated tasks.
    */
-  private class ExecutorAllocationListener extends SparkListener {
+  private[spark] class ExecutorAllocationListener extends SparkListener {
 
     private val stageIdToNumTasks = new mutable.HashMap[Int, Int]
     private val stageIdToTaskIndices = new mutable.HashMap[Int, mutable.HashSet[Int]]
@@ -733,20 +437,14 @@ private[spark] class ExecutorAllocationManager(
     val sourceName = "ExecutorAllocationManager"
     val metricRegistry = new MetricRegistry()
 
-    private def registerGauge[T](name: String, value: => T, defaultValue: T): Unit = {
-      metricRegistry.register(MetricRegistry.name("executors", name), new Gauge[T] {
-        override def getValue: T = synchronized { Option(value).getOrElse(defaultValue) }
+    def registerGauge[T](metric: Metric[T]): Unit = {
+      metricRegistry.register(MetricRegistry.name("executors", metric.name), new Gauge[T] {
+        override def getValue: T = synchronized {
+          Option(metric.getValue).getOrElse(metric.defaultValue)
+        }
       })
     }
-
-    registerGauge("numberExecutorsToAdd", numExecutorsToAdd, 0)
-    registerGauge("numberExecutorsPendingToRemove", executorsPendingToRemove.size, 0)
-    registerGauge("numberAllExecutors", executorIds.size, 0)
-    registerGauge("numberTargetExecutors", numExecutorsTarget, 0)
-    registerGauge("numberMaxNeededExecutors", maxNumExecutorsNeeded(), 0)
+    registerGauge(new Metric("numberAllExecutors", executorIds.size, 0))
+    allocationStrategy.getMetricsToRegister().foreach(metric => registerGauge(metric))
   }
-}
-
-private object ExecutorAllocationManager {
-  val NOT_SET = Long.MaxValue
 }
